refactor: decompose task-execution into focused modules
Refactor the monolithic 850-line task-execution.ts into a clean multi-module architecture with separate concerns:
- prepare-sandbox: Sandbox creation
- setup-git: Git authentication, identity, clone, and setup
- connect-assistant: Provider auth and OpenCode client setup
- run-prompt: Prompt execution with credential refresh
- stream-events: Real-time SSE event streaming
- helpers: Shared utilities and task database updates

The top-level index.ts now reads as a simple, readable sequence of logical steps using established patterns (Promise.all for parallel operations).
